+++ pycharm/src/main/java/rw/handler/runConf/RunConfHandlerManager.java
@@ -17,13 +17,10 @@ import org.jetbrains.annotations.VisibleForTesting;
 import rw.audit.RwSentry;
 import rw.consts.Const;
 import rw.pkg.BuiltinPackageManager;
-import rw.pkg.WebPackageManager;
 import rw.service.Service;
 import rw.util.OsType;
 
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 
 public class RunConfHandlerManager {
@@ -31,15 +28,17 @@ public class RunConfHandlerManager {
     @VisibleForTesting
     public BuiltinPackageManager builtinPackageManager;
     @VisibleForTesting
-    public WebPackageManager webPackageManager;
 
     public static RunConfHandlerManager singleton = null;
 
     Map<ExecutionEnvironment, BaseRunConfHandler> all;
+    @Nullable
+    BaseRunConfHandler last;
 
     @VisibleForTesting
     public RunConfHandlerManager() {
         this.all = new HashMap<>();
+        this.last = null;
     }
 
     public static RunConfHandlerManager get() {
@@ -52,6 +51,7 @@ public class RunConfHandlerManager {
 
     public void register(ExecutionEnvironment executionEnvironment, BaseRunConfHandler baseRunConfHandler) {
         this.all.put(executionEnvironment, baseRunConfHandler);
+        this.last = baseRunConfHandler;
     }
 
     @Nullable
@@ -59,6 +59,29 @@ public class RunConfHandlerManager {
         return this.all.get(executionId);
     }
 
+    public List<BaseRunConfHandler> getForProject(Project project) {
+        List<BaseRunConfHandler> ret = new ArrayList<>();
+
+        this.all.forEach((key, value) -> {
+            if(value.getProject() == project) {
+                ret.add(value);
+            }
+        });
+        return ret;
+    }
+
+    public List<BaseRunConfHandler> getAllActiveHandlers(@Nullable Project project) {
+        List<BaseRunConfHandler> ret = new ArrayList<>();
+
+        this.all.forEach((key, value) -> {
+            if((project == null || value.getProject() == project) && value.isActive()) {
+                ret.add(value);
+            }
+        });
+        return ret;
+    }
+
+    @Nullable
     public BaseRunConfHandler getCurrentHandler(Project project) {
         XDebugSessionImpl debugSession = ((XDebugSessionImpl) XDebuggerManager.getInstance(project).getCurrentSession());
 
@@ -74,6 +97,11 @@ public class RunConfHandlerManager {
 
         BaseRunConfHandler handler = this.all.get(environment);
         return handler;
+    }
+
+    @Nullable
+    public BaseRunConfHandler getLastHandler() {
+        return this.last;
     }
 
     public void deactivateAll() {