+++ pycharm/CHANGELOG.rst
@@ -2,6 +2,61 @@ Change Log
 ##########
 
 
+0.8.8
+-------
+
+**Miscellaneous**:
+    
+  Fixed:
+    * Empty working directory issues
+    
+0.8.7
+-------
+
+**Miscellaneous**:
+    
+  Fixed:
+    * Fix adding from import items issue
+    * Fix windows multiprocessing bugs
+    
+**PyCharm**:
+    
+  Added:
+    * Easier downgrading
+    
+  Removed:
+    * Package autoupdater
+    
+  Fixed:
+    * Confusing RELOADIUMPATH message when no files are watched
+    * Remote interpreter issues for new PyCharm versions
+    * View pane null pointer exception
+    
+0.8.6
+-------
+
+**Miscellaneous**:
+    
+  Fixed:
+    * Encoding issues
+    * Import threading issues
+    * Multiprocessing issues
+    * Double popup issue on FrameError
+    
+  Added:
+    * Multiprocessing extension
+    * Manual reload file command
+    * Symlinks and mounted directories support
+    
+**PyCharm**:
+    
+  Added:
+    * Reloadable files highlighting
+    * Manual reload action
+    
+  Fixed:
+    * Too many open files issue
+    
 0.8.5
 -------